init commit - must redraw for new view?
