+++ src/components/ConditionalLink.js
@@ -1,0 +1,7 @@
+import { Link } from 'react-router-dom';
+
+const ConditionalLink = ({ children, to, condition }) => (!!condition && to)
+      ? <Link to={to}>{children}</Link>
+      : <>{children}</>;
+
+export default ConditionalLink;
